refactor/#365: MavpView Wrapper 분리
